feat: Add support for executing queries and working with the raw TupleVectors produced
