+++ payloadbuilder-core/src/main/java/se/kuseman/payloadbuilder/core/BaseQueryResult.java
@@ -1,0 +1,7 @@
+package se.kuseman.payloadbuilder.core;
+
+/** Base interface of a query result. */
+public interface BaseQueryResult
+{
+    boolean hasMoreResults();
+}
